+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Fix 500 on favicon in admin [#126](https://github.com/etalab/udata-front/pull/126)
 - Update search results [#110](https://github.com/etalab/udata-front/pull/110) [#134](https://github.com/etalab/udata-front/pull/134)
 - Fix test error with long reuse title [#133](https://github.com/etalab/udata-front/pull/133)
+- Removed manifest logic [#129](https://github.com/etalab/udata-front/pull/129)
 
 ## 2.0.5 (2022-06-14)